clh: Hotplug the network devices
Instead of creating the VM with the network devices already plugged in,
let's actually add the network device *after* the VM is created, but
before the Vm is actually booted.

Although it looks like it doesn't make any difference of what we're
doing of what we used to do, this will be used to workaround a
limitation on openAPI when it comes to what we pass down to Cloud
Hypervisor so it can open the tuntap device.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>